+++ flexmeasures/api/v3_0/tests/conftest.py
@@ -5,6 +5,7 @@ from datetime import timedelta
 import pandas as pd
 import pytest
 from flask_security import SQLAlchemySessionUserDatastore, hash_password
+from sqlalchemy import select, delete
 
 from flexmeasures import Sensor, Source, User, UserRole
 from flexmeasures.data.models.generic_assets import GenericAssetType, GenericAsset
@@ -20,7 +21,7 @@ def setup_api_test_data(
     """
     print("Setting up data for API v3.0 tests on %s" % db.engine)
     sensors = add_incineration_line(
-        db, User.query.get(setup_roles_users["Test Supplier User"])
+        db, db.session.get(User, setup_roles_users["Test Supplier User"])
     )
     return sensors
 
@@ -33,10 +34,11 @@ def setup_api_fresh_test_data(
     Set up fresh data for API dev tests.
     """
     print("Setting up fresh data for API 3.0 tests on %s" % fresh_db.engine)
-    for sensor in Sensor.query.all():
-        fresh_db.delete(sensor)
+    for sensor in fresh_db.session.scalars(select(Sensor)).all():
+        fresh_db.session.execute(delete(Sensor).filter_by(id=sensor.id))
     sensors = add_incineration_line(
-        fresh_db, User.query.get(setup_roles_users_fresh_db["Test Supplier User"])
+        fresh_db,
+        fresh_db.session.get(User, setup_roles_users_fresh_db["Test Supplier User"]),
     )
     return sensors
 
@@ -108,7 +110,8 @@ def add_asset_with_children(db, setup_roles_users):
         generic_asset_type=parent_type,
         account_id=test_supplier_user,
     )
-    db.session.flush()  # assign sensor ids
+    db.session.add(parent)
+    db.session.flush()  # assign parent asset id
 
     assets = [
         GenericAsset(
@@ -121,7 +124,7 @@ def add_asset_with_children(db, setup_roles_users):
     ]
 
     db.session.add_all(assets)
-    db.session.flush()  # assign sensor ids
+    db.session.flush()  # assign children asset ids
 
     assets.append(parent)