+++ flexmeasures/api/v3_0/tests/utils.py
@@ -1,4 +1,7 @@
+from sqlalchemy import select
+
 from flexmeasures import Sensor
+from flexmeasures.data import db
 
 
 def make_sensor_data_request_for_gas_sensor(
@@ -10,7 +13,9 @@ def make_sensor_data_request_for_gas_sensor(
     """Creates request to post sensor data for a gas sensor.
     This particular gas sensor measures units of m³/h with a 10-minute resolution.
     """
-    sensor = Sensor.query.filter(Sensor.name == "some gas sensor").one_or_none()
+    sensor = db.session.execute(
+        select(Sensor).filter_by(name="some gas sensor")
+    ).scalar_one_or_none()
     values = num_values * [-11.28]
     if include_a_null:
         values[0] = None