+++ flexmeasures/data/services/data_sources.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from flask import current_app
+from sqlalchemy import select
 
 from flexmeasures import User
 from flexmeasures.data import db
@@ -18,7 +19,7 @@ def get_or_create_source(
 ) -> DataSource:
     if is_user(source):
         source_type = "user"
-    query = DataSource.query.filter(DataSource.type == source_type)
+    query = select(DataSource).filter(DataSource.type == source_type)
     if model is not None:
         query = query.filter(DataSource.model == model)
     if version is not None:
@@ -33,7 +34,7 @@ def get_or_create_source(
         query = query.filter(DataSource.name == source)
     else:
         raise TypeError("source should be of type User or str")
-    _source = query.one_or_none()
+    _source = db.session.execute(query).scalar_one_or_none()
     if not _source:
         if is_user(source):
             _source = DataSource(user=source, model=model, version=version)
@@ -62,8 +63,8 @@ def get_source_or_none(
     :param source:      source id
     :param source_type: optionally, filter by source type
     """
-    query = DataSource.query
+    query = select(DataSource)
     if source_type is not None:
         query = query.filter(DataSource.type == source_type)
     query = query.filter(DataSource.id == int(source))
-    return query.one_or_none()
+    return db.session.execute(query).scalar_one_or_none()